Deleting all the classes that we are not using anymore
Added a new Classes called LeagueController in order to create inside and inside of teams we created players

Also created a Configuration file in order to use a profile to populated the league, I didn't created a default profile because when there is no profile, the default profile is setted already.

Created three models, Player, Team and League, where League is the main class, and in there I created the object Team and in Team created Players

To initialize the league I created a method and added the annotation @PostContruct so when the programs start it will populate the league,team and player, therefore there is a config file with the profile to do that,

Tested some endpoint therefore I will continue testing it because need more time to complete the excersice.

i tested with postman sending link and datas by url.