Refactorizing more files such as classes, i finally deleted leagueService and leagueConfiguration, this because I don't use it anymore.
The LeagueRepository was making some issues so I decided to initialize everything in the service in order to maintain everything a little bit more explicit

Created the rest of the endpoints and tested with postman

Added a .zip file where there is the json file with the post, gets, put and delete request for testing.

I tried to maintain the leagueController as cleaner as I can.

Created cleared messages with log.

Adding a postContruct annotation in basket and soccer leagues in order to initilize the leagues using the profile "populated"

Keep working on DTO's